requirements.txt
blis==0.4.1
boto3==1.12.26
botocore==1.15.26
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
cymem==2.0.3
docutils==0.15.2
gensim==3.8.1
google-api-core==1.16.0
google-auth==1.11.3
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
importlib-metadata==1.5.0
jmespath==0.9.5
murmurhash==1.0.2
numpy==1.18.2
pandas==1.0.3
plac==1.1.3
preshed==3.0.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyspellchecker==0.5.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
scipy==1.4.1
six==1.14.0
smart-open==1.10.0
spacy==2.2.4
spacy-spanish-lemmatizer==0.3
srsly==1.0.2
thinc==7.4.0
tqdm==4.43.0
urllib3==1.25.8
wasabi==0.6.0
zipp==3.1.0
